# packages/frontend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "德州扑克游戏API",
    "version": "1.0.0",
    "description": "德州扑克游戏后端API文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/tournaments": {
      "post": {
        "summary": "创建新的锦标赛",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "锦标赛名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "锦标赛描述"
                  },
                  "buy_in": {
                    "type": "number",
                    "description": "买入费用"
                  },
                  "max_players": {
                    "type": "number",
                    "description": "最大玩家数量"
                  },
                  "start_time": {
                    "type": "string",
                    "format": "date-time",
                    "description": "开始时间"
                  },
                  "blind_structure_id": {
                    "type": "string",
                    "description": "盲注结构ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "锦标赛创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tournament"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "get": {
        "summary": "获取锦标赛列表",
        "tags": [
          "Tournaments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "registration",
                "in_progress",
                "completed",
                "cancelled"
              ]
            },
            "description": "锦标赛状态"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number"
            },
            "description": "限制返回数量"
          }
        ],
        "responses": {
          "200": {
            "description": "锦标赛列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tournament"
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/{id}": {
      "get": {
        "summary": "获取锦标赛详情",
        "tags": [
          "Tournaments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "锦标赛ID"
          }
        ],
        "responses": {
          "200": {
            "description": "锦标赛详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tournament"
                }
              }
            }
          },
          "404": {
            "description": "锦标赛不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/{id}/register": {
      "post": {
        "summary": "注册参加锦标赛",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "锦标赛ID"
          }
        ],
        "responses": {
          "201": {
            "description": "注册成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "锦标赛不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "delete": {
        "summary": "取消锦标赛注册",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "锦标赛ID"
          }
        ],
        "responses": {
          "200": {
            "description": "取消注册成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "注册信息不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/{id}/start": {
      "post": {
        "summary": "开始锦标赛",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "锦标赛ID"
          }
        ],
        "responses": {
          "200": {
            "description": "锦标赛开始成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "锦标赛不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/user": {
      "get": {
        "summary": "获取用户注册的锦标赛",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "用户注册的锦标赛列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tournament"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/{id}/rankings": {
      "get": {
        "summary": "获取锦标赛排名",
        "tags": [
          "Tournaments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "锦标赛ID"
          }
        ],
        "responses": {
          "200": {
            "description": "锦标赛排名列表"
          },
          "404": {
            "description": "锦标赛不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/tournaments/default-blind-structure": {
      "post": {
        "summary": "创建默认盲注结构",
        "tags": [
          "Tournaments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "默认盲注结构创建成功"
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  },
...
}